pyproject.toml
[project]
name = "legalai"
version = "0.1.0"
description = "Simple NLP application for analyzing legal documents using CanLII API"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "python-dotenv>=1.0.1,<2.0.0",
    "requests>=2.31.0,<3.0.0",
    "spacy>=3.7.2,<4.0.0",
    "beautifulsoup4>=4.12.2,<5.0.0",
    "openai>=1.12.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.4.0",
    "sentence-transformers>=2.5.0",
    "torch>=2.2.0"
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
